- refactors animation provider to be a global wrapper instead of music page
