+++ zune_ui/lib/pages/music_page/music_categories/index.dart
@@ -7,6 +7,7 @@ import 'package:memoized/memoized.dart';
 import 'package:provider/provider.dart';
 import 'package:zune_ui/enums/index.dart';
 import 'package:zune_ui/providers/global_state/index.dart';
+import 'package:zune_ui/providers/animation_provider/index.dart';
 import 'package:zune_ui/widgets/common/index.dart';
 
 /// NOTE: Scoping imports behind parent, so that console log is exposed from Music Page
