+++ zune_ui/lib/pages/music_page/view_selector/index.dart
@@ -10,6 +10,7 @@ import 'package:zune_ui/pages/music_page/genre_list/index.dart';
 import 'package:zune_ui/pages/music_page/playlist_list/index.dart';
 import 'package:zune_ui/pages/music_page/track_list/index.dart';
 import 'package:zune_ui/providers/global_state/index.dart';
+import 'package:zune_ui/providers/animation_provider/index.dart';
 
 /// NOTE: Scoping imports behind parent, so that console log is exposed from Music Page
 import 'package:zune_ui/pages/music_page/index.dart' as parent;
