+++ zune_ui/lib/providers/animation_provider/index.dart
@@ -1,0 +1,9 @@
+library animation_provider;
+
+import 'package:zune_ui/widgets/custom/debug_print.dart';
+
+import 'package:flutter/widgets.dart';
+
+part "animation_provider.dart";
+
+final console = DebugPrint().register(DebugComponent.animation);
